# 8276162: Optimise unsigned comparison pattern