@TobiHartmann 2021-11-17
This introduced a regression on 32-bit x86, @merykitty could you please have a look?
[JDK-8277324](https://bugs.openjdk.java.net/browse/JDK-8277324)