@TobiHartmann 2021-11-17
The problem is missing match rules on x86-32: https://github.com/openjdk/jdk/pull/6427