@rose00 2024-10-17
More completely, it is x +-^ Integer.MIN_VALUE
That is, add, sub, xor are all legitimate idioms for flipping the sign